test(frontdesk): split autosync_test.go by concern
At 4254 lines it was more than twice the test ceiling. The 110 tests
split along the same seams the production code took in #786, and the
shared harness gets its own file:

- autosync_harness_test.go: stubAutoMember and its recorders, the fleet
  builders (incompleteFleet, hashFleet), enableAutoSync, the small state
  readers and the shared fixtures
- autosync_hold_test.go: the sync hold (schema/build skew, release)
- autosync_incomplete_test.go: incomplete applies, the retry rate limit,
  the edge-triggered event
- autosync_divergence_test.go: flagging, the unreadable-hash clock, the
  recovery edge, message wording
- autosync_test.go keeps the loop itself (coalescing, converge, force,
  rearm, the push/verify stamps)

Pure moves; no test body changes. The entry leaves the size allowlist.
