+++ scripts/ci/size-allowlist.txt
@@ -10,7 +10,6 @@
 internal/api/applogs_test.go	2083
 internal/api/discovery_integration_quota_test.go	2735
 internal/failover/circuitbreaker_test.go	2056
-internal/frontdesk/autosync_test.go	4254
 internal/model/model_crud_test.go	2399
 internal/provider/discovery_test.go	2235
 internal/proxy/model_gone_test.go	2739
